[Refactor] JWT 발급 시 memberId 함께 반환 (#71)
* [Refactor] JWT 발급시 memberId 함께 반환

* [Fix] 깨진 테스트 수정

+++ src/main/java/com/parting/dippin/api/auth/service/AuthenticationService.java
@@ -11,11 +11,11 @@ public class AuthenticationService {
 
     private final TokenProvider tokenProvider;
 
-    public GetJwtResDto login(long memberId) {
+    public GetJwtResDto login(int memberId) {
         return tokenProvider.createJwt(memberId);
     }
 
-    public GetJwtResDto reissue(long memberId) {
+    public GetJwtResDto reissue(int memberId) {
         return tokenProvider.createJwt(memberId);
     }
 }
